requirements.txt
defusedxml==0.7.1
numpy==1.19.5
pandas==1.2.4
Pillow==8.2.0
sklearn==0.0
tensorflow==2.5.0
wai.annotations==0.7.1
wai.annotations.voc==1.1.0
wai.common==0.0.36
